--- Assets/UniGLTF/Runtime/UniGLTF/IO/AnimationImporter.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: d602384685dd4f179350052013659720
-timeCreated: 1537445972
