+++ Assets/UniGLTF/Runtime/UniGLTF/IO/AnimationImporterUtil.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a6e858fe43dba6342902d72392b1bdce
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
