+++ Assets/UniGLTF/Runtime/UniGLTF/IO/IAnimationImporter.cs
@@ -1,0 +1,7 @@
+﻿namespace UniGLTF
+{
+    public interface IAnimationImporter
+    {
+        void Import(ImporterContext context);
+    }
+}
