+++ Assets/UniGLTF/Runtime/UniGLTF/IO/IAnimationImporter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4f950ba9271cbff4cbc4345f5a23a5d8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
