+++ Assets/UniGLTF/Runtime/UniGLTF/IO/RootAnimationImporter.cs
@@ -1,0 +1,10 @@
+﻿namespace UniGLTF
+{
+    public sealed class RootAnimationImporter : IAnimationImporter
+    {
+        public void Import(ImporterContext context)
+        {
+            AnimationImporterUtil.ImportAnimation(context);
+        }
+    }
+}
