+++ Assets/UniGLTF/Runtime/UniGLTF/IO/RootAnimationImporter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 27640e6274339664ea492827be5a2217
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
